Move VoltageInitializer to network.util package
